+++ src/View/OptionsParser.php
@@ -102,9 +102,12 @@ class OptionsParser
      * </code>
      * @param string ...$classes Classes as string, or multiple arguments
      * @return $this
+     * @deprecated 2.25.0 Use instead `AddButtonClassesTrait::addButtonClasses()`
      */
     public function addButtonClasses(string ...$classes)
     {
+        deprecationWarning('Deprecated. Use instead `AddButtonClassesTrait::addButtonClasses()`');
+
         $baseClasses = ['primary', 'secondary', 'success', 'danger', 'warning', 'info', 'light', 'dark', 'link'];
         $allClasses = [...$baseClasses, 'outline-primary', 'outline-secondary', 'outline-success', 'outline-danger', 'outline-warning', 'outline-info', 'outline-light', 'outline-dark', 'lg', 'sm', 'block'];